Adding the release for result report E3.2
While not being a classical software release, this commit marks the end
of the first iteration of the control node development in the
DevOpt-project. The executable jar-File in this archive along with the
additional files and directories is executable on Windows and Linux
(Ubuntu) with a compliance level of Java version 11. However, the
capabilities are rather limited yet as described in detail in result
report E3.2.