Addition of ant build script and corresponding release
This build script eases local testing of an executable Jar-archive
(deployable instance) of a control node. Hence, it creates a
Zip-archive, which contains that Jar build from the current state of the
source files and includes all model and configuration files required to
execute the control node with default settings.