DevOpt meta model update
Container restriction regarding the mandatory availability of an
identification description results in not accepting multiple entries for
that container. The model is simply not loaded during tests, but without
any error, warning, etc.
For now, exclude this restriction. Later, we will update the meta model
anyway due to recent project results.