EASy resource management update
In order to avoid correct placing of the startup file and the DevOpt
meta model files for starting the EASy-Producer components, this update
introduces self-management. All resources required to start EASy are now
available as resources internally. These resource will be extracted to
files into a runtime directory and delete again when EASy is stopped.
User do not have to care anymore about these runtime files.