Internal model information management updated
Instead of using a simple list of available EntityInfo instances, this
update introduces a HashMap, which enables the usage of keys to identify
these instances. Hence, not entire instances must be distributed to
other components, but only the key and the data from the EntityInfo that
the respective component really needs. For later mapping back, the
components have the keys introduced with this update.