Establishing monitoring also for available models
During setup, the model manager loads all models from the model
directory to create the respective EntityInfo instances. Thats just the
initialization. However, monitoring of the respective entities must be
established as well (not only, if a new registration is received).
Hence, when run() is called, for each loaded model, monitoring will be
established before the ModelReceiver instance is started.