Fix monitoring scope parsing with multiple colons
If a valid monitoring scope definition is given, like
"devopttestmonitoring@tcp://broker.hivemq.com:1883", parsing that string
to create an EntityInfo instance produced a wrong URL due to not
considering multiple colons in the string. This is now fixed by using
lastIndexOf() when searching for the colon. According tests are added
and all tests are executed successfully.