Model unloading added
The IVML models are only used to create EntityInfo instances. Hence, the
models can be unloaded again. This update introduces this unloading for
both, successful as well as unsuccessful model additions. For the
latter, this update removes any create information until the point the
addition fails. This includes that even correctly created EntityInfo
instances will be deleted, if establishing their monitoring fails.