Consider multiple models for configuration loading
If multiple IVML projects with the same name, but in different files,
are added, loading the corresponding configuration for the file, which
was added last, yields the wrong configuration. This update now
considers all available models with a specific project name and checks
in addition, if the model information has the correct IVML file to
ensure that the correct configuration for a given IVML file is loaded.