Addition of Aggregator
This update introduces the first basic version of the aggregator. It
uses the configuration parameter of the setup as added by the previous
commits. However, the aggregation algorithm is hard-coded and specific
to the physical demonstrator in the DevOpt project. This must be
improved later. Further, only basic setup tests exists. More must
follow. All tests executed successfully.