Correct MQTT client usage
IDK why it was introduced, but recurrently connecting and disconnecting
is not how MQTT works. This update fixes that issue, which caused
problems in some scenario tests outside this project. In essence,
publishing and subscribing only connects the client once, if it is not
already connected. Disconnecting the client again is an explicit call
now, which typically has to be done by the parent instance using the
MQTT client.
All tests executed successfully (without changes).