Delete model manager instance after tests
Later introduction of scenario tests require that all global instances
will be created again. Hence, the current tests must ensure that the
instances will be deleted after tests are done. This was not the case
for the model manager tests, where also the AllTest class was in charge
of deleting some resources exclusive to these tests. This update moves
resource management to the abstract model manager test class, uses the
logger there instead of System.out directly, and ensures that after a
class, the model manager instance will be deleted again.