Monitoring data receiver becomes its own thread
In order to decouple reception of monitoring messages and their actual
processing, the monitoring data receiver now runs in its own thread.
This update is the first step towards a total decoupling, which will be
completed by the next updates. However, the introduction of threading to
this receiver resulted in some fundamental changes, which affect a
larger set of other classes as well. Keeping these changes in individual
commits was not possible without committing defective code.
All tests executed successfully.