Clean creation, start, and stop of monitoring data receiver
After messing with the mess in previous commits, this update finally
introduces the correct instance creation, start, stop, and deletion of
the monitoring data receiver. With this update, it is safe to stop and
(re-)start that instance. This is also reflected in the new abstract
test class all classes must extend that rely on the monitoring data
receiver (except for the ModelManager tests, which already extend other
classes).
All tests executed successfully.