Additional, separate model generator
In order ease the generation of IVML models, which describe entities for
tests and later also the demonstrator entities, this update introduces a
utility class, which supports such generation. The models are simple and
not very flexible, but for the current state of the implementation it is
suitable.
The generator may be extended in future to create more complex models
and to provide more customization options.
Due to this explicit generator class, the abstract scenario test class
does not need to provide the same functionality twice, but uses that new
generator, which in turn impacts the calls from the specific scenario
classes.
All tests executed successfully.