Network connection setup
This update improves the validation of different network connections,
e.g. for registration or aggregation. Further, it improves the
description of the respective configuration properties and error
messages. As a result, aggregation as an optional connection must be
defined either completely or not at all leading to a termination and
setup exception, if defined incompletely. The applies to incomplete or
ambiguous network definitions for registration.
According changes to existing tests and addition of new setup tests.
All tests executed successfully.